Delete User 2

+++ sql/user_list.php
@@ -15,7 +15,7 @@ if ($users){
         $user_list .= "<td>{$user['last_name']} </td>";
         $user_list .= "<td>{$user['last_login']} </td>";
         $user_list .= "<td><a href=\"modify-user.php?user_id={$user['id']}\">Edit</a> </td>";
-        $user_list .= "<td><a href=\"delete-user.php?user_id={$user['id']}\">Delete</a> </td>";
+        $user_list .= "<td><a href=\"delete-user.php?user_id={$user['id']}\" onclick=\"return confirm('Are You Sure');\">Delete</a> </td>";
         $user_list .= "</tr>";
     }
 }else{
